Created a calculator that takes the sides of two triangles as input and displays if the triangles are congruent or not
